server/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "prompt-catalog-mcp"
version = "1.0.0"
description = "MCP server and CLI for the Prompt Catalog — an open-source prompt library for AI-assisted software development"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [{ name = "Prompt Catalog Contributors" }]
keywords = ["mcp", "prompt", "ai", "llm", "software-development", "sdlc"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
]

dependencies = [
    "mcp>=1.0.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "click>=8.0",
    "jsonschema>=4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
prompt-catalog = "prompt_catalog_mcp.cli:main"

[project.urls]
Homepage = "https://github.com/KevinRabun/prompt-catalog"
Repository = "https://github.com/KevinRabun/prompt-catalog"
Issues = "https://github.com/KevinRabun/prompt-catalog/issues"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
